- Rename db to driver - Grouped redis db drivers into delegate module
