+++ cilantro/__init__.py
@@ -4,4 +4,5 @@ conf = resource_filename(__name__, 'cilantro.conf')
 config = configparser.ConfigParser()
 config.read(conf)
 
-print(config['default'].get('masternode_ip'), config['default'].get('delegate_ip'), config['default'].get('witness_ip'))
+# TODO -- investigate why config['default'] key does not exist (line below is giving runtime error)
+# print(config['default'].get('masternode_ip'), config['default'].get('delegate_ip'), config['default'].get('witness_ip'))
