+++ cilantro/db/delegate/driver_base.py
@@ -1,5 +1,5 @@
 import redis
 
-class BaseDB(object):
+class DriverBase(object):
     def __init__(self, host='localhost', port=6379, db=0):
         self.r = redis.StrictRedis(host=host, port=port, db=db)
